test uav matching

+++ Block_wise_GPS_correction.py
@@ -2540,7 +2540,7 @@ class Patch:
 		cv2.waitKey(0)
 
 		img = green_channel-0.61*blue_channel-0.39*red_channel
-
+		img = img.astype('uint8')
 		min_p = np.amin(img)
 		max_p = np.amax(img)
 		rng = (max_p-min_p)
